Fix undefined obj bug in JS
When calling the 3 bar charts in violations.js together, or the 3
sunburst charts in sunburst.js together, there would be an undefined
object bug as all the related data are not ready yet for gem gon. So,
the anonymous function at the start of each js file was commented out.
Rather, the plotting functions are called separately from the
corresponding *.html.erb files.